views.py: delete(): added a condition to check user is valid; a messages.info for edit was added

+++ apps/core/views.py
@@ -8,12 +8,7 @@ import json
 from geopy.geocoders import MapBox
 import os
 
-from django.contrib.messages import constants as message_constants
-MESSAGE_TAGS = {message_constants.DEBUG: 'debug',
-                message_constants.INFO: 'info',
-                message_constants.SUCCESS: 'success',
-                message_constants.WARNING: 'warning',
-                message_constants.ERROR: 'danger',}
+
 
 class ReportForm(forms.ModelForm):
     class Meta:
@@ -97,7 +92,7 @@ def log_location(request):
 def edit(request, id):
     report = Report.objects.get(id=id)
 
-    if report.user == request.user:
+    if report.user != request.user:
         messages.error(request, 'Could not confirm if this record was filed \
                                     by the current user.')
         return redirect('/')
@@ -106,6 +101,7 @@ def edit(request, id):
         form = ReportForm(request.POST, request.FILES, instance=report)
         if form.is_valid():
             form.save()
+            messages.info(request, 'Your update was saved!')
             return redirect('/account/users/' + request.user.username)
         else:
             messages.error(request, 'Form Validation Error at the server')
@@ -129,6 +125,12 @@ def list_view(request):
 @login_required
 def delete(request, id):
     report = Report.objects.get(id=id)
+
+    if report.user != request.user:
+        messages.error(request, 'Could not confirm if this record was filed \
+                                    by the current user.')
+        return redirect('/')
+
     report.delete() 
     messages.warning(request, f"Deleted the report of \'{report.type}\'")
